requirements.txt
Pyrogram
TgCrypto
git+https://github.com/AsmSafone/mega.py.git
moviepy
filetype
ffmpeg-python
hurry.filesize
